txhash: move AuthHash into txhash
