+++ crates/core/txhash/src/lib.rs
@@ -1,5 +1,7 @@
 mod effect_hash;
 mod effecting_data;
+mod auth_hash;
 
 pub use effect_hash::EffectHash;
 pub use effecting_data::EffectingData;
+pub use auth_hash::{AuthHash, AuthorizingData};
